Extract package sleeper.clients.admin.properties
