+++ java/clients/src/main/java/sleeper/clients/admin/AdminCommonPrompts.java
@@ -15,6 +15,7 @@
  */
 package sleeper.clients.admin;
 
+import sleeper.clients.admin.properties.AdminClientPropertiesStore;
 import sleeper.clients.util.console.ConsoleInput;
 import sleeper.clients.util.console.ConsoleOutput;
 import sleeper.clients.util.console.menu.ConsoleChoice;
