+++ java/clients/src/main/java/sleeper/clients/admin/FilesStatusReportScreen.java
@@ -16,6 +16,7 @@
 
 package sleeper.clients.admin;
 
+import sleeper.clients.admin.properties.AdminClientPropertiesStore;
 import sleeper.clients.status.report.FilesStatusReport;
 import sleeper.clients.status.report.filestatus.StandardFileStatusReporter;
 import sleeper.clients.util.console.ConsoleInput;
