+++ java/clients/src/main/java/sleeper/clients/admin/TableNamesReport.java
@@ -15,6 +15,7 @@
  */
 package sleeper.clients.admin;
 
+import sleeper.clients.admin.properties.AdminClientPropertiesStore;
 import sleeper.clients.util.console.ConsoleInput;
 import sleeper.clients.util.console.ConsoleOutput;
