+++ java/clients/src/main/java/sleeper/clients/admin/properties/PropertyGroupWithCategory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.admin;
+package sleeper.clients.admin.properties;
 
 import sleeper.configuration.properties.PropertyGroup;
