+++ java/clients/src/test/java/sleeper/clients/admin/testutils/AdminClientMockStoreBase.java
@@ -18,7 +18,7 @@ package sleeper.clients.admin.testutils;
 import org.mockito.InOrder;
 import org.mockito.Mockito;
 
-import sleeper.clients.admin.AdminClientPropertiesStore;
+import sleeper.clients.admin.properties.AdminClientPropertiesStore;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.statestore.StateStore;
