+++ java/clients/src/test/java/sleeper/clients/admin/testutils/AdminConfigStoreTestHarness.java
@@ -15,7 +15,7 @@
  */
 package sleeper.clients.admin.testutils;
 
-import sleeper.clients.admin.AdminClientPropertiesStore;
+import sleeper.clients.admin.properties.AdminClientPropertiesStore;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
