+++ java/clients/src/test/java/sleeper/clients/admin/testutils/RunAdminClient.java
@@ -18,7 +18,7 @@ package sleeper.clients.admin.testutils;
 
 import sleeper.clients.AdminClient;
 import sleeper.clients.admin.AdminClientStatusStoreFactory;
-import sleeper.clients.admin.UpdatePropertiesWithTextEditor;
+import sleeper.clients.admin.properties.UpdatePropertiesWithTextEditor;
 import sleeper.clients.testutil.TestConsoleInput;
 import sleeper.clients.testutil.ToStringPrintStream;
 import sleeper.compaction.job.CompactionJobStatusStore;
